Segment responsibilities for card header
