+++ package.json
@@ -87,6 +87,7 @@
     "typescript": "^4.7.4"
   },
   "dependencies": {
+    "classnames": "^2.3.2",
     "lodash": "^4.17.5",
     "prop-types": "^15.5.10",
     "react-final-form": "^6.3.0",
