+++ src/components/CriteriaCard.tsx
@@ -1,21 +1,17 @@
-import {
-  Card,
-  IconButton,
-  Select,
-  SelectOptionType,
-} from '@folio/stripes/components';
-import React, { useEffect, useMemo } from 'react';
-import { Field, useField, useForm } from 'react-final-form';
+import { Card, IconButton } from '@folio/stripes/components';
+import React, { useMemo } from 'react';
+import { useField } from 'react-final-form';
 import {
   CriteriaCardGroupType,
   CriteriaGroup,
   CriteriaTerminal,
-  FormValues,
 } from '../form/types';
 
+import classNames from 'classnames';
+import { useFieldArray } from 'react-final-form-arrays';
 import css from './CriteriaCard.module.css';
 import CriteriaCardAddButton from './CriteriaCardAddButton';
-import { useFieldArray } from 'react-final-form-arrays';
+import CriteriaCardGroupSelect from './CriteriaCardGroupSelect';
 
 export interface CriteriaCardProps {
   prefix: string;
@@ -28,51 +24,10 @@ export default function CriteriaCard({
   root = false,
   onRemove,
 }: CriteriaCardProps) {
-  const selectDefaultValue = useMemo(() => {
-    if (root) {
-      return CriteriaCardGroupType.PASS;
-    } else {
-      return CriteriaCardGroupType.ALL_OF;
-    }
-  }, [root]);
-
-  const form = useForm();
   const type = useField<CriteriaCardGroupType>(`${prefix}type`).input.value;
   const criteria = useFieldArray<CriteriaGroup | CriteriaTerminal>(
     `${prefix}criteria`
   );
-
-  const selectOptions = useMemo(() => {
-    const options: SelectOptionType<CriteriaCardGroupType>[] = [
-      {
-        label: 'All of:',
-        value: CriteriaCardGroupType.ALL_OF,
-      },
-      {
-        label: 'Any of:',
-        value: CriteriaCardGroupType.ANY_OF,
-      },
-      {
-        label: 'None of:',
-        value: CriteriaCardGroupType.NONE_OF,
-      },
-    ];
-
-    if (root) {
-      options.unshift({
-        label: 'No criteria (always run)',
-        value: CriteriaCardGroupType.PASS,
-      });
-    }
-
-    return options;
-  }, [root]);
-
-  useEffect(() => {
-    if (type === CriteriaCardGroupType.PASS && criteria.fields.length) {
-      form.change(`${prefix}criteria`, []);
-    }
-  }, [type]);
 
   const headerEnd = useMemo(() => {
     if (type === CriteriaCardGroupType.PASS) {
@@ -94,21 +49,13 @@ export default function CriteriaCard({
   return (
     <Card
       headerClass={css.headerClass}
-      headerStart={
-        <Field name={`${prefix}type`} defaultValue={selectDefaultValue}>
-          {(fieldProps) => (
-            <Select<CriteriaCardGroupType>
-              {...fieldProps}
-              required
-              marginBottom0
-              dataOptions={selectOptions}
-            />
-          )}
-        </Field>
-      }
+      headerStart={<CriteriaCardGroupSelect prefix={prefix} root={root} />}
       headerEnd={headerEnd}
+      cardClass={classNames({
+        [css.emptyPassCard]: type === CriteriaCardGroupType.PASS,
+      })}
     >
-      card contents
+      <div></div>
     </Card>
   );
 }
